checkIfInMyCorner implementiert und getestet mit Hashfunktion auf true- und false-Werte Corner und Node von Integer zu Double geändert

+++ app/src/main/java/com/example/joshi/can/Logic/Corner.java
@@ -8,10 +8,10 @@ import com.example.joshi.can.Exception.YMustBeLargerThanZeroException;
  */
 
 public class Corner {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Corner(int x,int y) throws XMustBeLargerThanZeroException, YMustBeLargerThanZeroException {
+    public Corner(double x,double y) throws XMustBeLargerThanZeroException, YMustBeLargerThanZeroException {
 
         if(x < 0){
             throw new XMustBeLargerThanZeroException("X has to be larger than Zero");
@@ -25,19 +25,19 @@ public class Corner {
 
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
